Refactor index.ts files to separate business logic
- Move nodejs-adapter logic to node-server.ts
- Move pino-adapter logic to pino-logger-factory.ts
- Move zod-openapi-plugin logic to zod-openapi-plugin.ts
- Rename converter.ts to zod-schema-converter.ts for clarity
- Update all index.ts files to only handle exports
- Add corresponding changesets for version management
