+++ .changeset/refactor-nodejs-adapter-index.md
@@ -1,0 +1,5 @@
+---
+'@korix/nodejs-adapter': patch
+---
+
+Refactor nodejs-adapter to move server logic out of index.ts
