+++ .changeset/refactor-pino-adapter-index.md
@@ -1,0 +1,5 @@
+---
+'@korix/pino-adapter': patch
+---
+
+Refactor pino-adapter to move logger factory logic out of index.ts
