+++ .changeset/refactor-zod-openapi-plugin.md
@@ -1,0 +1,5 @@
+---
+'@korix/zod-openapi-plugin': patch
+---
+
+Refactor zod-openapi-plugin to move logic out of index.ts and improve file naming.
