+++ packages/nodejs-adapter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @korix/nodejs-adapter
 
+## 0.1.0-alpha.3
+
+### Patch Changes
+
+- Refactor nodejs-adapter to move server logic out of index.ts
+
 ## 0.1.0-alpha.2
 
 ### Patch Changes
