+++ packages/nodejs-adapter/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@korix/nodejs-adapter",
-  "version": "0.1.0-alpha.2",
+  "version": "0.1.0-alpha.3",
   "description": "Node.js adapter for Kori framework",
   "type": "module",
   "author": "mitz (@bufferings)",
